nar/reader.go: use wire package, drop serialize*.go
